seperated dashboard functions to new file
