+++ client/components/Dashboard.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 
-import { dashboardTab } from '../actions'
+import { changeTab } from './component_functions/change_tab'
 
 import MySharingDetails from '../containers/MySharingDetails.jsx'
 import MyProfile from '../containers/MyProfile.jsx'
@@ -25,17 +25,6 @@ function mapStateToProps (state) {
     dispatch: state.dispatch,
     dashboardState: state.dashboardState
   }
-}
-
-function changeTab (props, clickedTab) {
-  if (props.dashboardState === 'myProfile') {
-    document.getElementById('sharing-button').classList.add('dashboard-active-button')
-    document.getElementById('profile-button').classList.remove('dashboard-active-button')
-  } else {
-    document.getElementById('sharing-button').classList.remove('dashboard-active-button')
-    document.getElementById('profile-button').classList.add('dashboard-active-button')
-  }
-  props.dispatch(dashboardTab(clickedTab))
 }
 
 export default connect(mapStateToProps)(Dashboard)
